Moved sign-components to their own comp-folder.
