+++ src/app/(authentication)/sign-out/page.tsx
@@ -1,5 +1,5 @@
 import {FC} from "react";
-import SignOutCC from "@/components/ui/SignOutCC";
+import SignOutCC from "@/components/ui/authentication/SignOutCC";
 
 const page: FC = () => {
     return <SignOutCC/>;
